Revert MapSet as plugin
