+++ src/core/immerClass.ts
@@ -23,9 +23,7 @@ import {
 	getCurrentScope,
 	NOTHING,
 	freeze,
-	current,
-	proxyMap,
-	proxySet
+	current
 } from "../internal"
 
 interface ProducersFns {
@@ -204,9 +202,9 @@ export function createProxy<T extends Objectish>(
 ): Drafted<T, ImmerState> {
 	// precondition: createProxy should be guarded by isDraftable, so we know we can safely draft
 	const draft: Drafted = isMap(value)
-		? proxyMap(value, parent)
+		? getPlugin("MapSet").proxyMap_(value, parent)
 		: isSet(value)
-		? proxySet(value, parent)
+		? getPlugin("MapSet").proxySet_(value, parent)
 		: createProxyProxy(value, parent)
 
 	const scope = parent ? parent.scope_ : getCurrentScope()
