+++ src/internal.ts
@@ -9,4 +9,3 @@ export * from "./core/finalize"
 export * from "./core/proxy"
 export * from "./core/immerClass"
 export * from "./core/current"
-export * from "./core/mapset"
