+++ src/types/index.js.flow
@@ -105,6 +105,7 @@ declare export function createDraft<T>(base: T): T
  */
 declare export function finishDraft<T>(base: T, listener?: PatchListener): T
 
+declare export function enableMapSet(): void
 declare export function enablePatches(): void
 
 declare export function freeze<T>(obj: T, freeze?: boolean): T
