+++ tsconfig.json
@@ -13,5 +13,5 @@
     "module": "ES6",
     "allowJs": true
   },
-  "files": ["./src/immer.ts", "./src/plugins/patches.ts", "src/types/globals.d.ts"]
+  "files": ["./src/immer.ts", "./src/plugins/mapset.ts", "./src/plugins/patches.ts", "src/types/globals.d.ts"]
 }
