+++ website/docs/installation.mdx
@@ -30,9 +30,46 @@ The following features can be opt-in to:
 
 | Feature | Description | Method to call |
 | --- | --- | --- |
+| [ES2015 Map and Set support](./complex-objects.md) | To enable Immer to operate on the native `Map` and `Set` collections, enable this feature | `enableMapSet()` |
 | [JSON Patch support](./patches.mdx) | Immer can keep track of all the changes you make to draft objects. This can be useful for communicating changes using JSON patches | `enablePatches()` |
 
-Vanilla Immer kicks in at ~5KB gzipped. Every plugin that is enabled adds < 1 KB to that. The breakdown is as follows:
+For example, if you want to use `produce` on a `Map`, you need to enable this feature once during the start of your application:
+
+```typescript
+// In your application's entrypoint
+import {enableMapSet} from "immer"
+
+enableMapSet()
+
+// ...later
+import {produce} from "immer"
+
+const usersById_v1 = new Map([
+	["michel", {name: "Michel Weststrate", country: "NL"}]
+])
+
+const usersById_v2 = produce(usersById_v1, draft => {
+	draft.get("michel").country = "UK"
+})
+
+expect(usersById_v1.get("michel").country).toBe("NL")
+expect(usersById_v2.get("michel").country).toBe("UK")
+```
+
+Vanilla Immer kicks in at ~3KB gzipped. Every plugin that is enabled adds < 1 KB to that. The breakdown is as follows:
+
+```
+Import size report for immer:
+┌───────────────────────┬───────────┬────────────┬───────────┐
+│        (index)        │ just this │ cumulative │ increment │
+├───────────────────────┼───────────┼────────────┼───────────┤
+│ import * from 'immer' │   5873    │     0      │     0     │
+│        produce        │   3421    │    3421    │     0     │
+│     enableMapSet      │   4837    │    4845    │   1424    │
+│     enablePatches     │   4240    │    5655    │    810    │
+└───────────────────────┴───────────┴────────────┴───────────┘
+(this report was generated by npmjs.com/package/import-size)
+```
 
 ## Immer on older JavaScript environments?
